Added Profile page styled accordingly to the prototype
